primeiro refatoramento
