+++ chatApp/tools.go
@@ -1,0 +1,10 @@
+package chatApp
+
+func StringInSlice(a string, list []string) bool {
+	for _, b := range list {
+		if b == a {
+			return true
+		}
+	}
+	return false
+}
